+++ packages/cloud-scheduler-utils/src/index.ts
@@ -1,5 +1,8 @@
 export { createHttpJob } from './create-jobs.js'
+export type { CreateHttpJobConfig } from './create-jobs.js'
 
 export { deleteAllJobs, deleteJob } from './delete-jobs.js'
+export type { DeleteAllJobsConfig, DeleteJobConfig } from './delete-jobs.js'
 
 export { randomSchedule } from './random-schedule.js'
+export type { Config as RandomScheduleConfig } from './random-schedule.js'